\name{ddPROC}
\alias{ddPROC}
\docType{data}
\title{ Processed miRNA data  (RGList) }
\description{
	Filtered and Normalized miRNA data stored in a RGList object. 
}
\usage{data(ddPROC)}
\details{
  ddPROC is originated after the processing of the dd.micro raw data. 
}
\author{ Pedro Lopez-Romero }
\seealso{
	An overview of how ddPROC is obtained is given in
  	\code{filterMicroRna}
}
\keyword{datasets}
